requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pdfplumber==0.9.0
pymupdf==1.23.8
python-docx==0.8.11
openpyxl==3.1.2
